refactor: move is-defined
