+++ src/core/util/index.ts
@@ -4,6 +4,5 @@ export * from './create-processor.fn';
 export * from './is-array-of-buildables';
 export * from './is-buildable';
 export * from './is-builder.fn';
-export * from './is-defined';
 export * from './is-processor.fn';
 export * from './unwrap-if-buildable';
