+++ src/core/util/is-buildable.ts
@@ -1,5 +1,5 @@
+import { isDefined } from '../../util';
 import { Buildable, BuildableSymbol } from '../types';
-import { isDefined } from './is-defined';
 
 export function isBuildable(value: any): value is Buildable<any> {
   return isDefined(value) && isDefined(value[BuildableSymbol]);
