+++ src/core/util/is-builder.fn.ts
@@ -1,5 +1,5 @@
+import { isDefined } from '../../util';
 import { BuilderFnSymbol } from '../types';
-import { isDefined } from './is-defined';
 
 export function isBuilderFunction(fn: Function): boolean {
   return isDefined(fn) && isDefined((fn as any)[BuilderFnSymbol]);
