+++ src/core/util/is-processor.fn.ts
@@ -1,5 +1,5 @@
+import { isDefined } from '../../util';
 import { ProcessorSymbol, ProcessorType } from '../types';
-import { isDefined } from './is-defined';
 
 export function isProcessorFn(fn: Function, ofType?: ProcessorType): boolean {
   if (isDefined(ofType)) {
